+++ api-gateway/src/main/java/n11bootcamp_project_backend/api_gateway/ApiGatewayApplication.java
@@ -3,9 +3,13 @@ package n11bootcamp_project_backend.api_gateway;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.web.bind.annotation.CrossOrigin;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {
+		"n11bootcamp_project_backend"
+})
 @EnableDiscoveryClient
+@CrossOrigin
 public class ApiGatewayApplication {
 
 	public static void main(String[] args) {